refactor: nits and comments
