+++ test/entry.ts
@@ -236,37 +236,17 @@ describe('Entry', () => {
     });
 
     it('should convert a string timestamp', () => {
-      const tests = [
-        {
-          inputTime: '2020-01-01T00:00:00.11Z',
-          expectedSeconds: 1577836800,
-          expectedNanos: 110000000,
-        },
-        {
-          inputTime: '2020-01-01T00:00:00Z',
-          expectedSeconds: 1577836800,
-          expectedNanos: 0,
-        },
-        {
-          inputTime: '2020-01-01T00:00:00.999999999Z',
-          expectedSeconds: 1577836800,
-          expectedNanos: 999999999,
-        },
-        {
-          inputTime: 'invalid timestamp string',
-          expectedSeconds: 0,
-          expectedNanos: 0,
-        },
-      ];
-
-      for (const test of tests) {
-        entry.metadata.timestamp = test.inputTime;
-        const json = entry.toJSON();
-        assert.deepStrictEqual(json.timestamp, {
-          seconds: test.expectedSeconds,
-          nanos: test.expectedNanos,
-        });
-      }
+      const test = {
+        inputTime: '2020-01-01T00:00:00.999999999Z',
+        expectedSeconds: 1577836800,
+        expectedNanos: 999999999,
+      };
+      entry.metadata.timestamp = test.inputTime;
+      const json = entry.toJSON();
+      assert.deepStrictEqual(json.timestamp, {
+        seconds: test.expectedSeconds,
+        nanos: test.expectedNanos,
+      });
     });
 
     it('should convert a raw incoming HTTP request', () => {
